+++ src/main/java/com/nhom7/coworkingspace/exception/EmailAlreadyExistsException.java
@@ -1,0 +1,12 @@
+package com.nhom7.coworkingspace.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.CONFLICT)
+public class EmailAlreadyExistsException extends RuntimeException {
+
+    public EmailAlreadyExistsException() {
+        super("Email is already in use");
+    }
+}