+++ src/main/java/com/nhom7/coworkingspace/service/AuthService.java
@@ -1,0 +1,8 @@
+package com.nhom7.coworkingspace.service;
+
+import com.nhom7.coworkingspace.dto.request.SignUpRequest;
+
+public interface AuthService {
+
+    void signUp(SignUpRequest request);
+}