+++ src/main/java/com/nhom7/coworkingspace/service/OtpService.java
@@ -3,4 +3,6 @@ package com.nhom7.coworkingspace.service;
 public interface OtpService {
 
     void sendConfirmationOtp(String email);
+
+    void sendPasswordResetOtp(String email);
 }